pthread_once: use cfg_if to select Linux or MacOS version
